test(backend): Add poll.service.spec.ts
